Cyber_Project/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
cryptography==41.0.7
Werkzeug==3.0.1
email-validator==2.1.0
pyotp==2.9.0
qrcode[pil]==7.4.2
Pillow==10.0.0
